Move platform specific stuff out of const
